removed warnings flag on clippy from ci. updated duplicate cast in merge solution.

+++ src/top_interview_150/merge_sorted_arrays.rs
@@ -63,10 +63,10 @@ impl Solution {
 
         for i in (0..nums1.len()).rev() {
             if m < 0 || n >= 0 && nums1[m as usize] < nums2[n as usize] {
-                nums1[i as usize] = nums2[n as usize];
+                nums1[i] = nums2[n as usize];
                 n -= 1;
             } else {
-                nums1[i as usize] = nums1[m as usize];
+                nums1[i] = nums1[m as usize];
                 m -= 1;
             }
         }
